# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank Card Management API
  description: "API for managing bank cards, transfers, and user administration"
  contact:
    name: Support
    email: support@bank.com
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- Bearer Authentication: []
tags:
- name: Card Management
  description: Endpoints for managing user cards
- name: Admin Management
  description: Endpoints for administrators
paths:
  /api/cards:
    get:
      tags:
      - Card Management
      summary: Get user cards
      description: Returns paginated list of user's cards with optional filtering
      operationId: getMyCards
      parameters:
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCardResponse'
    post:
      tags:
      - Card Management
      summary: Create a new card
      description: Creates a new bank card for the authenticated user
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
        required: true
      responses:
        "400":
          description: Invalid input
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
        "201":
          description: Card created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
  /api/cards/{cardId}/block:
    post:
      tags:
      - Card Management
      summary: Block a card
      description: "Blocks a specific card, preventing further transactions"
      operationId: blockCard
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Card not found
        "403":
          description: You don't own this card
        "200":
          description: Card blocked successfully
  /api/cards/transfer:
    post:
      tags:
      - Card Management
      summary: Transfer between cards
      description: Transfers money between two cards owned by the same user
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
        required: true
      responses:
        "400":
          description: Invalid amount or insufficient funds
        "200":
          description: Transfer completed successfully
        "404":
          description: Card not found
  /api/auth/register:
    post:
      tags:
      - auth-controller
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
  /api/auth/login:
    post:
      tags:
      - auth-controller
      operationId: authenticateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/admin/cards/{cardId}/activate:
    post:
      tags:
      - Admin Management
      summary: Activate a card
      description: Activates a blocked card
      operationId: activateCard
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/cards/{cardId}/balance:
    get:
      tags:
      - Card Management
      summary: Get card balance
      description: Returns the current balance of a specific card
      operationId: getBalance
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: number
  /api/admin/users:
    get:
      tags:
      - Admin Management
      summary: Get all users
      description: Returns list of all users in the system
      operationId: getAllUsers
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageUser'
  /api/admin/cards:
    get:
      tags:
      - Admin Management
      summary: Get all cards
      description: Returns paginated list of all cards in the system
      operationId: getAllCards
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCardResponse'
  /api/admin/users/{userId}:
    delete:
      tags:
      - Admin Management
      summary: Delete a user
      description: Permanently deletes a user from the system
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/admin/cards/{cardId}:
    delete:
      tags:
      - Admin Management
      summary: Delete a card
      description: Permanently deletes a card from the system
      operationId: deleteCard
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
components:
  schemas:
    CardRequest:
      type: object
      properties:
        initialBalance:
          minimum: 0
          type: number
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedNumber:
          type: string
        ownerName:
          type: string
        expiryDate:
          type: string
          format: date
        status:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
        balance:
          type: number
    TransferRequest:
      required:
      - amount
      - fromCardId
      - toCardId
      type: object
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          minimum: 1
          type: number
    RegisterRequest:
      required:
      - email
      - fullName
      - password
      - username
      type: object
      properties:
        username:
          maxLength: 50
          minLength: 3
          type: string
        password:
          maxLength: 100
          minLength: 6
          type: string
        email:
          type: string
        fullName:
          type: string
    Card:
      type: object
      properties:
        id:
          type: integer
          format: int64
        encryptedCardNumber:
          type: string
        maskedNumber:
          type: string
        owner:
          $ref: '#/components/schemas/User'
        expiryDate:
          type: string
          format: date
        status:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
        balance:
          type: number
        createdAt:
          type: string
          format: date-time
    GrantedAuthority:
      type: object
      properties:
        authority:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        password:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum:
          - ROLE_USER
          - ROLE_ADMIN
        createdAt:
          type: string
          format: date-time
        cards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
        active:
          type: boolean
        enabled:
          type: boolean
        accountNonLocked:
          type: boolean
        accountNonExpired:
          type: boolean
        credentialsNonExpired:
          type: boolean
        authorities:
          type: array
          items:
            $ref: '#/components/schemas/GrantedAuthority'
    AuthRequest:
      required:
      - password
      - username
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
        username:
          type: string
        role:
          type: string
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageCardResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardResponse'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
    PageUser:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/User'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
  securitySchemes:
    Bearer Authentication:
      type: http
      name: Bearer Authentication
      scheme: bearer
      bearerFormat: JWT